se verifica que los libros borrados pertenecen al usuario

+++ app/PHP/mod_libros/borrar_libros.php
@@ -9,13 +9,37 @@
       die("Database connection failed: " . $conn->connect_error);
     }
 
+    if (isset($_COOKIE["username"])) { // TODO: Verificar la sesión
+      $user = $_COOKIE["username"];
+    } else {
+      die("No puedes borrar libros sin haber iniciado sesión");
+    }
+    // Obtener la lista de libros escritos por este usuario, para asegurarse que solo se borran los escritos por él
+    $query = mysqli_prepare($conn, "SELECT `Book ID` FROM escritos WHERE `Used ID`=?") or die (mysqli_error($conn));
+    mysqli_stmt_bind_param($query, "s", $us_id);
+    $us_id = $user;
+    mysqli_stmt_execute($query) or die (mysqli_error($conn));
+    $libros_usuario = mysqli_stmt_get_result($query)->fetch_array();
+
     $libros = $_POST["libros_borrados"];
     $arrlength = count($libros);
     for($x = 0; $x < $arrlength; $x++) {
-      $query = mysqli_prepare($conn, "DELETE FROM libro WHERE `Book ID`=?") or die (mysqli_error($conn));
-      mysqli_stmt_bind_param($query, "s", $lib_id);
-      $lib_id = $libros[$x];
-      mysqli_stmt_execute($query) or die (mysqli_error($conn));
+      $hostname = "db";
+      $username = "admin";
+      $password = "test";
+      $db = "database";
+    
+      $conn = mysqli_connect($hostname,$username,$password,$db);
+      if ($conn->connect_error) {
+        die("Database connection failed: " . $conn->connect_error);
+      }
+
+      if (in_array($libros[$x], $libros_usuario)) {
+        $query = mysqli_prepare($conn, "DELETE FROM libro WHERE `Book ID`=?") or die (mysqli_error($conn));
+        mysqli_stmt_bind_param($query, "s", $lib_id);
+        $lib_id = $libros[$x];
+        mysqli_stmt_execute($query) or die (mysqli_error($conn));
+      }
     }
